Removes Test Shutdown
This test was not in the spirit of the python wrapper tests, as it went
too 'deep' into what it was testing.  The goal of the wrapper tests is
to test that the api calls are exposed, and that to some extend, the
objects are handled correctly.  They are not intended to test the
functionality of the libvcx library.

Signed-off-by: Mark Hadley <mark.hadley@evernym.com>

Former-commit-id: f86c3d113848ab81678f7f82ddf771c9b54cabd0